% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goric_penalty.R
\name{goric_penalty}
\alias{goric_penalty}
\alias{orglm_penalty}
\title{GORIC penalty term}
\usage{
goric_penalty(object, iter = 1e+05, type = "GORIC", mc.cores = 1)

orglm_penalty(object, iter = 1e+05, type = "GORIC", mc.cores = 1)
}
\arguments{
\item{object}{an object of class orlm, orgls (or orglm for function \code{orglm_penalty})}

\item{iter}{number of iterations to calculate GORIC penalty terms}

\item{type}{if \code{"GORIC"} (default), the penalty term for the generalized order restriction information criterion is computed; with \code{"GORICCa"} or \code{"GORICCb"} small sample corrections for the penalty term are applied}

\item{mc.cores}{number of cores using a socket cluster implemented in package \code{parallel}}
}
\description{
Calculates the GORIC penalty term (level probability) by Monte-Carlo simulation.
}
\seealso{
\code{\link{orlm}}, \code{\link{orgls}}, \code{\link{orglm}}
}
\keyword{misc}
