\name{orlmcontrol}
\alias{orlmcontrol}
\title{Control arguments for the orlm function.}
\description{A list with control arguments controlling the orlm function}
\usage{
orlmcontrol(maxiter = 10000, absval = 1e-04)
}
\arguments{
  \item{maxiter}{maximum number of iterations}
  \item{absval}{tolerance criterion for convergence}
}
\value{a list with control arguments}

\author{Daniel Gerhard and Rebecca M. Kuiper}

\keyword{misc}
