\name{ormodel_methods}
\alias{summary.orlm}
\alias{print.orlm}
\alias{summary.orgls}
\alias{print.orgls}

\title{Summarize orlm Objects}
\description{\code{summary} method for class "orlm"}
\usage{
\method{summary}{orlm}(object, digits = max(3, getOption("digits")
- 3), ...)
\method{print}{orlm}(x, digits = max(3, getOption("digits")
- 3), ...)
}
\arguments{
  \item{object}{an object of class "orlm"}
  \item{x}{an object of class "orlm"}
  \item{digits}{the number of significant digits to use when printing.}
  \item{...}{further arguments passed to or from other methods}
}

\author{Daniel Gerhard and Rebecca M. Kuiper}

\keyword{methods}
