% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orlmcontrol.R
\name{orlmcontrol}
\alias{orlmcontrol}
\title{Control arguments for the orlm function.}
\usage{
orlmcontrol(maxiter = 10000, absval = 1e-04)
}
\arguments{
\item{maxiter}{maximum number of iterations}

\item{absval}{tolerance criterion for convergence}
}
\value{
a list with control arguments
}
\description{
A list with control arguments controlling the orlm function
}
\seealso{
\code{\link{orlm}}
}
\keyword{models}
